move streaming/utils/transcript.py to streaming/models/transcript.py
As suggested by @ajar98
